Added: ProcessFolderAction and FavouritesAction
