+++ resources/lib/actions/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: CC-BY-NC-SA-4.0
 
 __all__ = ["addonaction", "channellistaction", "categoryaction", "vaultaction", "keyword",
-           "logaction", "configurechannelaction"]
+           "logaction", "configurechannelaction", "folderaction", "showfavsaction"]
