+++ resources/lib/plugin.py
@@ -4,7 +4,6 @@
 import os
 import xbmcplugin
 import xbmc
-import xbmcgui
 
 from resources.lib import envcontroller
 from resources.lib.logger import Logger
@@ -15,14 +14,10 @@ from resources.lib.xbmcwrapper import XbmcWrapper, XbmcDialogProgressWrapper, Xb
 from resources.lib.mediaitem import MediaItem
 from resources.lib.helpers.channelimporter import ChannelIndex
 from resources.lib.helpers.languagehelper import LanguageHelper
-from resources.lib.helpers.stopwatch import StopWatch
 from resources.lib.helpers.sessionhelper import SessionHelper
-from resources.lib.helpers.htmlentityhelper import HtmlEntityHelper
 from resources.lib.textures import TextureHandler
 from resources.lib.paramparser import ParameterParser
 from resources.lib.urihandler import UriHandler
-from resources.lib.channelinfo import ChannelInfo
-from resources.lib.chn_class import Channel
 from resources.lib.actions import keyword
 
 
@@ -103,7 +98,7 @@ class Plugin(ParameterParser):
             env_ctrl.cache_clean_up(Config.cacheDir, Config.cacheValidTime)
 
             # empty picklestore
-            self._pickler.purge_store(Config.addonId)
+            self.pickler.purge_store(Config.addonId)
 
         # create a session
         SessionHelper.create_session(Logger.instance())
@@ -199,14 +194,17 @@ class Plugin(ParameterParser):
 
             elif self.params[keyword.ACTION] == self.actionFavourites:
                 # we should show the favourites
-                self.show_favourites(self.channelObject)
+                from resources.lib.actions.favouritesaction import ShowFavouritesAction
+                addon_action = ShowFavouritesAction(self, self.channelObject)
 
             elif self.params[keyword.ACTION] == self.actionAllFavourites:
-                self.show_favourites(None)
+                from resources.lib.actions.favouritesaction import ShowFavouritesAction
+                addon_action = ShowFavouritesAction(self, None)
 
             elif self.params[keyword.ACTION] == self.actionListFolder:
                 # channelName and URL is present, Parse the folder
-                self.process_folder_list()
+                from resources.lib.actions.folderaction import FolderAction
+                addon_action = FolderAction(self, self.channelObject, self.media_item)
 
             elif self.params[keyword.ACTION] == self.actionPlayVideo:
                 self.play_video_item()
@@ -224,115 +222,6 @@ class Plugin(ParameterParser):
 
         self.__fetch_textures()
         return
-
-    def show_favourites(self, channel):
-        """ Show the favourites (for a channel).
-
-        :param ChannelInfo|None channel:    The channel to show favourites for.
-                                            Might be None to show all.
-
-        """
-
-        Logger.debug("Plugin::show_favourites")
-
-        if channel is None:
-            Logger.info("Showing all favourites")
-        else:
-            Logger.info("Showing favourites for: %s", channel)
-
-        # Local import for performance
-        from resources.lib.favourites import Favourites
-        f = Favourites(Config.favouriteDir)
-        favs = f.list(channel)
-        self.process_folder_list(favs)
-
-    def process_folder_list(self, favorites=None):
-        """Wraps the channel.process_folder_list
-
-        :param list[MediaItem]|None favorites:
-
-        """
-
-        Logger.info("Plugin::process_folder_list Doing process_folder_list")
-        try:
-            ok = True
-
-            # read the item from the parameters
-            selected_item = self.media_item
-
-            # determine the parent guid
-            parent_guid = self._get_parent_guid(self.channelObject, selected_item)
-
-            if favorites is None:
-                watcher = StopWatch("Plugin process_folder_list", Logger.instance())
-                media_items = self.channelObject.process_folder_list(selected_item)
-                watcher.lap("Class process_folder_list finished")
-            else:
-                watcher = StopWatch("Plugin process_folder_list With Items", Logger.instance())
-                media_items = favorites
-
-            if len(media_items) == 0:
-                Logger.warning("process_folder_list returned %s items", len(media_items))
-                ok = self.__show_empty_information(media_items, favs=favorites is not None)
-            else:
-                Logger.debug("process_folder_list returned %s items", len(media_items))
-
-            kodi_items = []
-
-            for media_item in media_items:  # type: MediaItem
-                self.__update_artwork(media_item, self.channelObject)
-
-                if media_item.type == 'folder' or media_item.type == 'append' or media_item.type == "page":
-                    action = self.actionListFolder
-                    folder = True
-                elif media_item.is_playable():
-                    action = self.actionPlayVideo
-                    folder = False
-                else:
-                    Logger.critical("Plugin::process_folder_list: Cannot determine what to add")
-                    continue
-
-                # Get the Kodi item
-                kodi_item = media_item.get_kodi_item()
-                self.__set_kodi_properties(kodi_item, media_item, folder,
-                                           is_favourite=favorites is not None)
-
-                # Get the context menu items
-                context_menu_items = self.__get_context_menu_items(self.channelObject, item=media_item)
-                kodi_item.addContextMenuItems(context_menu_items)
-
-                # Get the action URL
-                url = media_item.actionUrl
-                if url is None:
-                    url = self.create_action_url(self.channelObject, action=action, item=media_item, store_id=parent_guid)
-
-                # Add them to the list of Kodi items
-                kodi_items.append((url, kodi_item, folder))
-
-            watcher.lap("Kodi Items generated")
-
-            # add items but if OK was False, keep it like that
-            ok = ok and xbmcplugin.addDirectoryItems(self.handle, kodi_items, len(kodi_items))
-            watcher.lap("items send to Kodi")
-
-            if ok and parent_guid is not None:
-                self._pickler.store_media_items(parent_guid, selected_item, media_items)
-
-            watcher.stop()
-
-            self.__add_sort_method_to_handle(self.handle, media_items)
-            self.__add_breadcrumb(self.handle, self.channelObject, selected_item)
-
-            # set the content. It needs to be "episodes" to make the MediaItem.set_season_info() work
-            xbmcplugin.setContent(handle=self.handle, content="episodes")
-
-            xbmcplugin.endOfDirectory(self.handle, ok)
-        except Exception:
-            Logger.error("Plugin::Error Processing FolderList", exc_info=True)
-            XbmcWrapper.show_notification(LanguageHelper.get_localized_string(LanguageHelper.ErrorId),
-                                          LanguageHelper.get_localized_string(LanguageHelper.ErrorList),
-                                          XbmcWrapper.Error, 4000)
-            xbmcplugin.endOfDirectory(self.handle, False)
 
     # @LockWithDialog(logger=Logger.instance())  No longer needed as Kodi will do this automatically
     def play_video_item(self):
@@ -461,203 +350,6 @@ class Plugin(ParameterParser):
                     w.close()
         return
 
-    def __add_sort_method_to_handle(self, handle, items=None):
-        """ Add a sort method to the plugin output. It takes the Add-On settings into
-        account. But if none of the items have a date, it is forced to sort by name.
-
-        :param int handle:              The handle to add the sortmethod to.
-        :param list[MediaItem] items:   The items that need to be sorted
-
-        :rtype: None
-
-        """
-
-        if AddonSettings.mix_folders_and_videos():
-            label_sort_method = xbmcplugin.SORT_METHOD_LABEL_IGNORE_FOLDERS
-        else:
-            label_sort_method = xbmcplugin.SORT_METHOD_LABEL
-
-        if items:
-            has_dates = len(list([i for i in items if i.has_date()])) > 0
-            if has_dates:
-                Logger.debug("Sorting method: Dates")
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_DATE)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=label_sort_method)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_TRACKNUM)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_UNSORTED)
-                return
-
-            has_tracks = len(list([i for i in items if i.has_track()])) > 0
-            if has_tracks:
-                Logger.debug("Sorting method: Tracks")
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_TRACKNUM)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_DATE)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=label_sort_method)
-                xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_UNSORTED)
-                return
-
-        Logger.debug("Sorting method: Default (Label)")
-        xbmcplugin.addSortMethod(handle=handle, sortMethod=label_sort_method)
-        xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_DATE)
-        xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_TRACKNUM)
-        xbmcplugin.addSortMethod(handle=handle, sortMethod=xbmcplugin.SORT_METHOD_UNSORTED)
-        return
-
-    def __get_context_menu_items(self, channel, item=None):
-        """ Retrieves the custom context menu items to display.
-
-        favouritesList : Boolean   - Indication that the menu is for the favorites
-
-        :param Channel|None channel:    The channel from which to get the context menu items.
-                                        The channel might be None in case of some actions that
-                                        do not require a channel.
-        :param MediaItem|None item:     The item to which the context menu belongs.
-
-        :return: A list of context menu names and their commands.
-        :rtype: list[tuple[str,str]]
-
-        """
-
-        context_menu_items = []
-
-        # Generic, none-Python menu items that would normally cause an unwanted reload of the
-        # Python interpreter instance within Kodi.
-        refresh = LanguageHelper.get_localized_string(LanguageHelper.RefreshListId)
-        context_menu_items.append((refresh, 'Container.Refresh()'))
-
-        if item is None:
-            return context_menu_items
-
-        # if it was a favourites list, don't add the channel methods as they might be from a different channel
-        if channel is None:
-            return context_menu_items
-
-        if item.has_info():
-            info_action = LanguageHelper.get_localized_string(LanguageHelper.ItemInfo)
-            context_menu_items.append((info_action, 'Action(info)'))
-
-        # now we process the other items
-        possible_methods = self.__get_members(channel)
-        # Logger.Debug(possible_methods)
-
-        for menu_item in channel.contextMenuItems:
-            # Logger.Debug(menu_item)
-            if menu_item.itemTypes is None or item.type in menu_item.itemTypes:
-                # We don't care for complete here!
-                # if menu_item.completeStatus == None or menu_item.completeStatus == item.complete:
-
-                # see if the method is available
-                method_available = False
-
-                for method in possible_methods:
-                    if method == menu_item.functionName:
-                        method_available = True
-                        # break from the method loop
-                        break
-
-                if not method_available:
-                    Logger.warning("No method for: %s", menu_item)
-                    continue
-
-                cmd_url = self.create_action_url(channel, action=menu_item.functionName, item=item)
-                cmd = "RunPlugin(%s)" % (cmd_url,)
-                title = "Retro: %s" % (menu_item.label,)
-                Logger.trace("Adding command: %s | %s", title, cmd)
-                context_menu_items.append((title, cmd))
-
-        return context_menu_items
-
-    def __get_members(self, channel):
-        """ Caches the inspect.getmembers(channel) or dir(channel) method for performance
-        matters.
-
-        :param Channel channel:     The channel from which to get the context menu items.
-                                    The channel might be None in case of some actions that
-                                    do not require a channel.
-
-        :return: A list of all methods in the channel.
-        :rtype: list[str]
-
-        """
-
-        if channel.guid not in self.methodContainer:
-            # Not working on all platforms
-            # self.methodContainer[channel.guid] = inspect.getmembers(channel)
-            self.methodContainer[channel.guid] = dir(channel)
-
-        return self.methodContainer[channel.guid]
-
-    def __show_empty_information(self, items, favs=False):
-        """ Adds an empty item to a list or just shows a message.
-        @type favs: boolean
-        @param items:
-
-        :param list[MediaItem] items:   The list of items.
-        :param bool favs:               Indicating that we are dealing with favourites.
-
-        :return: boolean indicating to report the listing as succes or not.
-        :rtype: ok
-
-        """
-
-        if favs:
-            title = LanguageHelper.get_localized_string(LanguageHelper.NoFavsId)
-        else:
-            title = LanguageHelper.get_localized_string(LanguageHelper.ErrorNoEpisodes)
-
-        behaviour = AddonSettings.get_empty_list_behaviour()
-
-        Logger.debug("Showing empty info for mode (favs=%s): [%s]", favs, behaviour)
-        if behaviour == "error":
-            # show error
-            ok = False
-        elif behaviour == "dummy" and not favs:
-            # We should add a dummy items, but not for favs
-            empty_list_item = MediaItem("- %s -" % (title.strip("."), ), "", type='video')
-            empty_list_item.dontGroup = True
-            empty_list_item.complete = True
-            # add funny stream here?
-            # part = empty_list_item.create_new_empty_media_part()
-            # for s, b in YouTube.get_streams_from_you_tube("", self.channelObject.proxy):
-            #     part.append_media_stream(s, b)
-
-            # if we add one, set OK to True
-            ok = True
-            items.append(empty_list_item)
-        else:
-            ok = True
-
-        XbmcWrapper.show_notification(LanguageHelper.get_localized_string(LanguageHelper.ErrorId),
-                                      title, XbmcWrapper.Error, 2500)
-        return ok
-
-    def __set_kodi_properties(self, kodi_item, media_item, is_folder, is_favourite):
-        """ Sets any Kodi related properties.
-
-        :param xbmcgui.ListItem kodi_item:  The Kodi list item.
-        :param MediaItem media_item:        The internal media item.
-        :param bool is_folder:              Is this a folder.
-        :param bool is_favourite:           Is this a favourite.
-
-        """
-
-        # Set the properties for the context menu add-on
-        kodi_item.setProperty(self.propertyRetrospect, "true")
-        kodi_item.setProperty(self.propertyRetrospectFolder
-                              if is_folder
-                              else self.propertyRetrospectVideo, "true")
-
-        if is_favourite:
-            kodi_item.setProperty(self.propertyRetrospectFavorite, "true")
-        elif media_item.isCloaked:
-            kodi_item.setProperty(self.propertyRetrospectCloaked, "true")
-
-        if self.channelObject and self.channelObject.adaptiveAddonSelectable:
-            kodi_item.setProperty(self.propertyRetrospectAdaptive, "true")
-
-        if self.channelObject and self.channelObject.hasSettings:
-            kodi_item.setProperty(self.propertyRetrospectChannelSetting, "true")
-
     def __show_warnings(self, media_item):
         """ Show playback warnings for this MediaItem
 
@@ -676,29 +368,6 @@ class Plugin(ParameterParser):
                 title = LanguageHelper.get_localized_string(LanguageHelper.PaidTitle)
                 message = LanguageHelper.get_localized_string(LanguageHelper.PaidText)
                 XbmcWrapper.show_dialog(title, message)
-
-    # noinspection PyUnusedLocal
-    def __add_breadcrumb(self, handle, channel, selected_item, last_only=False):
-        """ Updates the Kodi category with a breadcrumb to the current parent item
-
-        :param int handle:                      The Kodi file handle
-        :param ChannelInfo|Channel channel:     The channel to which the item belongs
-        :param MediaItem selected_item:         The item from which to show the breadcrumbs
-        :param bool last_only:                  Show only the last item
-
-        """
-
-        bread_crumb = None
-        if selected_item is not None:
-            bread_crumb = selected_item.name
-        elif self.channelObject is not None:
-            bread_crumb = channel.channelName
-
-        if not bread_crumb:
-            return
-
-        bread_crumb = HtmlEntityHelper.convert_html_entities(bread_crumb)
-        xbmcplugin.setPluginCategory(handle=handle, category=bread_crumb)
 
     def __append_kodi_play_list(self, kodi_items):
         # Get the current playlist
@@ -733,43 +402,3 @@ class Plugin(ParameterParser):
             play_list.add(current_play_list_items[i].getfilename(), current_play_list_items[i])
 
         return start_url
-
-    def __update_artwork(self, media_item, channel):
-        """ Updates the fanart and icon of a MediaItem if thoses are missing.
-
-        :param MediaItem media_item:    The item to update
-        :param Channel channel:         A possible selected channel
-
-        """
-
-        if media_item is None:
-            return
-
-        if channel:
-            # take the channel values
-            fallback_icon = self.channelObject.icon
-            fallback_thumb = self.channelObject.noImage
-            fallback_fanart = self.channelObject.fanart
-            parent_item = channel.parentItem
-        else:
-            # else the Retrospect ones
-            fallback_icon = Config.icon
-            fallback_thumb = Config.fanart
-            fallback_fanart = Config.fanart
-            parent_item = None
-
-        if parent_item is not None:
-            fallback_thumb = parent_item.thumb or fallback_thumb
-            fallback_fanart = parent_item.fanart or fallback_fanart
-
-        # keep it or use the fallback
-        media_item.icon = media_item.icon or fallback_icon
-        media_item.thumb = media_item.thumb or fallback_thumb
-        media_item.fanart = media_item.fanart or fallback_fanart
-
-        if AddonSettings.use_thumbs_as_fanart() and \
-                TextureHandler.instance().is_texture_or_empty(media_item.fanart) and \
-                not TextureHandler.instance().is_texture_or_empty(media_item.thumb):
-            media_item.fanart = media_item.thumb
-
-        return
